+++ src/main/java/nextstep/member/MemberService.java
@@ -1,19 +1,14 @@
 package nextstep.member;
 
-import nextstep.auth.JwtTokenProvider;
-import nextstep.auth.TokenRequest;
-import nextstep.auth.TokenResponse;
 import nextstep.support.LoginException;
 import org.springframework.stereotype.Service;
 
 @Service
 public class MemberService {
     private final MemberDao memberDao;
-    private final JwtTokenProvider jwtTokenProvider;
 
-    public MemberService(MemberDao memberDao, JwtTokenProvider jwtTokenProvider) {
+    public MemberService(MemberDao memberDao) {
         this.memberDao = memberDao;
-        this.jwtTokenProvider = jwtTokenProvider;
     }
 
     public Long create(MemberRequest memberRequest) {
@@ -22,16 +17,6 @@ public class MemberService {
 
     public Member findById(Long id) {
         return memberDao.findById(id);
-    }
-
-    public TokenResponse createToken(TokenRequest tokenRequest) {
-        Member member = findByUsername(tokenRequest.getUsername());
-
-        if (member.checkWrongPassword(tokenRequest.getPassword())) {
-            throw new LoginException();
-        }
-
-        return new TokenResponse(jwtTokenProvider.createToken(member.getId().toString()));
     }
 
     public Member findByUsername(String username) {
