+++ src/test/java/nextstep/auth/LoginE2ETest.java
@@ -1,38 +1,106 @@
 package nextstep.auth;
 
 import io.restassured.RestAssured;
-import nextstep.member.Member;
+import nextstep.member.MemberRequest;
+import nextstep.theme.ThemeRequest;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 public class LoginE2ETest {
+    public static final String USERNAME = "username";
+    public static final String PASSWORD = "password";
+    private String token;
+
+    @Autowired
+    JwtTokenProvider jwtTokenProvider;
+
+    @BeforeEach
+    void setUp() {
+        MemberRequest body = new MemberRequest(USERNAME, PASSWORD, "name", "010-1234-5678");
+        var memberResponse = RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(body)
+                .when().post("/members")
+                .then().log().all()
+                .statusCode(HttpStatus.CREATED.value())
+                .extract();
+
+        String[] memberLocation = memberResponse.header("Location").split("/");
+        Long memberId = Long.parseLong(memberLocation[memberLocation.length - 1]);
+        token = jwtTokenProvider.createToken(memberId + "");
+    }
 
     @DisplayName("토큰을 생성한다")
     @Test
-    public void createToken() {
-        Member member = new Member("user", "pw", "user", "010-0000-0000");
-        TokenRequest tokenRequest = new TokenRequest("user", "pw");
-
-        RestAssured
+    public void create() {
+        TokenRequest body = new TokenRequest(USERNAME, PASSWORD);
+        var response = RestAssured
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(member)
-                .when().post("/members")
-                .then().log().all()
-                .statusCode(HttpStatus.CREATED.value());
-
-        RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(tokenRequest)
+                .body(body)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
                 .when().post("/login/token")
                 .then().log().all()
-                .statusCode(HttpStatus.OK.value());
+                .statusCode(HttpStatus.OK.value())
+                .extract();
+
+        assertThat(response.as(TokenResponse.class)).isNotNull();
+    }
+
+    @DisplayName("테마 목록을 조회한다")
+    @Test
+    public void showThemes() {
+        createTheme();
+
+        var response = RestAssured
+                .given().log().all()
+                .param("date", "2022-08-11")
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+                .when().get("/themes")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract();
+        assertThat(response.jsonPath().getList(".").size()).isEqualTo(1);
+    }
+
+    @DisplayName("테마를 삭제한다")
+    @Test
+    void delete() {
+        Long id = createTheme();
+
+        var response = RestAssured
+                .given().log().all()
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+                .when().delete("/themes/" + id)
+                .then().log().all()
+                .extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    public Long createTheme() {
+        ThemeRequest body = new ThemeRequest("테마이름", "테마설명", 22000);
+        String location = RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(body)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
+                .when().post("/themes")
+                .then().log().all()
+                .statusCode(HttpStatus.CREATED.value())
+                .extract().header("Location");
+        return Long.parseLong(location.split("/")[2]);
     }
 }
